Move Handler into relational package
